Move webpack folder into packages to be able to reference local upscaler directory
