+++ package.json
@@ -15,7 +15,8 @@
     "bootstrap": "lerna bootstrap --hoist",
     "server": "yarn serve -p 8000 packages/upscalerjs/test/lib/webpack-bundler/dist/",
     "test:unit": "yarn lerna run test:unit",
-    "test:integration:browser": "jest test/integration --config test/jestconfig.json --detectOpenHandles"
+    "test:integration:browser": "jest test/integration --config test/jestconfig.json --detectOpenHandles",
+    "test:startTestServer": "node test/lib/startServer.js"
   },
   "pre-commit": [
     "lint",
