+++ packages/test-scaffolding/index.js
@@ -1,6 +1,6 @@
 import * as tfjs from '@tensorflow/tfjs';
-// import Upscaler from 'upscaler';
+import Upscaler from 'upscaler';
 import flower from './flower.png';
 window.tfjs = tfjs;
 window.flower = flower;
-// window.upscaler = new Upscaler();
+window.upscaler = new Upscaler();
