+++ packages/test-scaffolding/package.json
@@ -2,8 +2,10 @@
   "name": "webpack-bundler",
   "version": "1.0.0",
   "main": "index.js",
+  "workspaceDependencies": {
+    "upscaler": "0.8.22"
+	},
   "dependencies": {
-    "@tensorflow/tfjs": "^3.2.0",
-    "upscaler": "^0.8.1"
+    "@tensorflow/tfjs": "^3.2.0"
   }
 }
